% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_databox.R
\docType{data}
\name{anno_databox}
\alias{anno_databox}
\title{Annotation data}
\format{
list of meta data and pair data for organisms:
\describe{
...
}
}
\source{
\url{""}
}
\usage{
anno_databox
}
\description{
Annotation data for overrepresentation and enrichment analysis
}
\keyword{datasets}
