% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{boxplot_byF1}
\alias{boxplot_byF1}
\title{Box plot by category/factor F1}
\usage{
boxplot_byF1(METBObj, xCol=1, factorLv1)
}
\arguments{
\item{METBObj}{METBObj object contains list of data.}

\item{xCol}{a column number or column name indicating which variable/feature in the \code{x_input} data frame to be plotted.}

\item{factorLv1}{a vector of category/factor levels.}
}
\value{
ggplot object.
}
\description{
Provide box plot of a variable/feature by category/factor F1.
}
\examples{
#boxplot_byF1(METBObj, 1, METBObj$Y)
}
