% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{boxplot_byF1F2}
\alias{boxplot_byF1F2}
\title{Box plot by category/factor F1 and F2}
\usage{
boxplot_byF1F2(METBObj, xCol=1, factorLv1, factorLv2)
}
\arguments{
\item{METBObj}{METBObj object contains list of data.}

\item{xCol}{a column number or column name indicating which variable/feature in the \code{x_input} data frame to be plotted.}

\item{factorLv1}{a vector of F1 category/factor levels.}

\item{factorLv2}{a vector of F2 category/factor levels.}
}
\value{
ggplot object.
}
\description{
Provide box plot of a variable/feature by category/factor F1 and F2.
}
\examples{
#sugar_dt = set_input_obj(sugar, 1,2,5)
#boxplot_byF1F2(sugar_dt, 1, sugar_dt$Y, sugar_dt$inputdata$time)
}
