% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{boxplot_overview}
\alias{boxplot_overview}
\title{Box plot overview}
\usage{
boxplot_overview(METBObj, plotvar=TRUE, plot_title="")
}
\arguments{
\item{METBObj}{METBObj object contains list of data.}

\item{plotvar}{a logical variable indicating box plot of variables/features (TRUE) or samples (FALSE).}

\item{plot_title}{text indicating plot title.}
}
\value{
ggplot object.
}
\description{
Provide box plot.
}
\examples{
#boxplot_overview(METBObj)
}
