% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_serrf}
\alias{check_serrf}
\title{Check serrf input}
\usage{
check_serrf(p)
}
\arguments{
\item{p}{data.table of sample information}
}
\value{
text message
}
\description{
The function check input data for serrf normalization.
}
\details{
The function is used inside normalize_input_data_byqc.
}
\examples{
#check_serrf(samp_info)
}
