% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemclass_databox.R
\docType{data}
\name{chemclass_databox}
\alias{chemclass_databox}
\title{Chemical class annotation data}
\format{
list of meta data and pair data for organisms:
\describe{
Downloaded from HMDB version 5.0 (released 2021-09-15).
}
}
\source{
\url{""}
}
\usage{
chemclass_databox
}
\description{
Chemical class annotation data for overrepresentation and enrichment analysis
}
\keyword{datasets}
