% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_overrep_analyze.R
\name{comb_overrep_analyze}
\alias{comb_overrep_analyze}
\title{Perform integrated pathway overrepresentation analysis}
\usage{
comb_overrep_analyze(input_data, organism="hsa", size=3)
}
\arguments{
\item{input_data}{a data frame of input data. The 1st column contains list of entities and the 2nd indicates entity types (i.e. gene, protein, compound). See details below.}

\item{organism}{a string specifying organism code from KEGG database. Choose one from hsa (default), tdc.}

\item{size}{a number specifying the minimum number of members in each annotation term to be used in the analysis. Default = 3.}
}
\value{
a list of the following components:

enrichment = enrichment results.

annotations = a data frame of annotation pairs.

details = a list of analysis details: univsize, testMethod, pAdjusted, pCombined, inputsize, inputtype, numsets, minsize.
}
\description{
perform integrated pathway overrepresentation analysis using hypergeometric tests.
If there are more than one entity types, Fisher's method is used to combine p-values before adjusted.
}
\details{
Metabox uses KEGG ID (e.g.C12078) for compounds, UniProt entry (e.g.P0C9J6) for proteins, Ensembl (e.g.ENSG00000139618) for genes.
}
\examples{
#out = comb_overrep_analyze(fnanal_data$combined_data, organism = "hsa", size = 3)
}
\references{
Johnson NL., Kotz S., and Kemp AW. (1992) Univariate Discrete Distributions, Second Edition. New York: Wiley.

Fisher R. (1932) Statistical methods for research workers. Oliver and Boyd, Edinburgh.
}
\seealso{
\code{\link{phyper}}, \code{\link{p.adjust}},\code{\link{pchisq}}
}
\author{
Kwanjeera W \email{kwanjeera.wan@mahidol.ac.th}
}
