% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_statplot}
\alias{combine_statplot}
\title{Combined statistics plot}
\usage{
combine_statplot(x_data, y_data, x_cutoff=0, y_cutoff=0, ptsize=3, plot_title="")
}
\arguments{
\item{x_data}{a numerical vector or a data frame of statistical values from x analysis, see details.}

\item{y_data}{a numerical vector or a data frame of statistical values from y analysis, see details.}

\item{x_cutoff}{text indicating plot title.}

\item{y_cutoff}{text indicating plot title.}

\item{ptsize}{a number of geom_point size.}

\item{plot_title}{text indicating plot title.}
}
\value{
ggplot object.
}
\description{
Provide plot of outputs from 2 statistical analyses.
}
\details{
A plot of outputs from two statistical analyses will generate.
A numerical vector or a one-column data frame of statistical values (e.g. p-values or VIP)
from each statistical analysis (e.g. t-test or PLS-DA) is needed for x_data and y_data.
Significance of a variable is denoted by its statistical value less than x_cutoff or more than y_cutoff.
}
\examples{
#sugar_dt = set_input_obj(sugar, 1,2,5)
#uout=univ_analyze(sugar_dt)
#mout=multiv_analyze(sugar_dt, method="pls", scale="pareto")
#combine_statplot(uout$p_adj,mout$vip_val,x_cutoff = 0.05,y_cutoff = 1)
}
