% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound_databox.R
\docType{data}
\name{compound_databox}
\alias{compound_databox}
\title{Compound data}
\format{
dataframe of compounds:
\describe{
Downloaded from HMDB version 5.0 (released 2021-09-15).
}
}
\source{
\url{""}
}
\usage{
compound_databox
}
\description{
Compound meta data for overrepresentation and enrichment analysis
}
\keyword{datasets}
