% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_analyze.R
\name{correlation_analyze}
\alias{correlation_analyze}
\title{Correlation analysis}
\usage{
correlation_analyze(METBObj, y=NULL, method="pearson")
}
\arguments{
\item{METBObj}{METBObj object contains list of data.}

\item{y}{a numeric vector or matrix or data frame of other variables e.g. genes, proteins or compounds, with samples/subjects as row.names. See details.}

\item{method}{name of multivariate analysis method. Choose one from the list: pearson, spearman. Default is pearson.}
}
\value{
a list of the following components:

p_value = a matrix of p-values.

p_adj = a matrix of adjusted p-values.

corr_coef = a matrix of correlation coefficient.

corr_data = a data frame of analysis result containing coeff,p_value,p_adj.

details = a list of analysis details: testMethod, pAdjusted.
}
\description{
Perform correlation analysis.
}
\details{
If y is provided, it will be combined to \code{METBObj$X} and correlation will be calculated for all \code{METBObj$X} and y variables.
}
\examples{
#sugar_dt = set_input_obj(sugar, 1,2,5)
#out=correlation_analyze(sugar_dt)
}
\seealso{
\code{\link[Hmisc:rcorr]{Hmisc::rcorr()}}, \code{\link[stats:p.adjust]{stats::p.adjust()}}
}
\author{
Kwanjeera W \email{kwanjeera.wan@mahidol.ac.th}
}
