% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{corrplot_heatmap}
\alias{corrplot_heatmap}
\title{Correlation heatmap}
\usage{
corrplot_heatmap(corr_data, plot_title="")
}
\arguments{
\item{corr_data}{data frame result from \code{correlation_analyze()}.
The data frame must contain Var1,Var2,coeff,p_value,p_adj columns.}

\item{plot_title}{text indicating plot title.}
}
\value{
ggplot object.
}
\description{
Provide correlation heatmap.
}
\examples{
#sugar_dt = set_input_obj(sugar, 1,2,5)
#out=correlation_analyze(sugar_dt)
#corrplot_heatmap(data.frame(out$corr_data))
}
