% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{densityplot_overview}
\alias{densityplot_overview}
\title{Density plot}
\usage{
densityplot_overview(x_input, plotvar=TRUE, plot_title="")
}
\arguments{
\item{x_input}{data frame of variables/features. Variables are in columns and samples are in rows.}

\item{plotvar}{a logical variable indicating density plot of variables/features (TRUE) or samples (FALSE).}

\item{plot_title}{text indicating plot title.}
}
\value{
ggplot object.
}
\description{
Provide density overview.
}
\examples{
#densityplot_overview(METBObj$X)
}
