% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analyze.R
\name{enrichment_analyze}
\alias{enrichment_analyze}
\title{Perform enrichment analysis}
\usage{
enrichment_analyze(input_data, pcol=2, fccol=NULL, method="reporter",
nodetype="compound", settype="pathway", organism="hsa", size=3)
}
\arguments{
\item{input_data}{a data frame of input data containing list of entities, statistics (e.g. p-values, t-values or F-values) and directions (e.g. fold-changes). See details below.}

\item{pcol}{a number specifying the column number that contains statistical values (e.g. p-values, t-values or F-values). Default is 2.}

\item{fccol}{a number specifying the column number that contains direction values (e.g. fold-changes). Default is NULL and directionality classes will be ignored.}

\item{method}{a string specifying the enrichment analysis method. It can be one of methods from \pkg{\link{piano}} e.g. reporter (default), fisher, median, mean, stouffer. See \code{\link{runGSA}}.}

\item{nodetype}{a string specifying a node type. It can be one of compound (default), protein, gene.}

\item{settype}{a string specifying a set type. It can be one of pathway (default), chemicalclass.}

\item{organism}{a string specifying organism code from KEGG database, the parameter will not be used for \code{settype = "chemicalclass"}. Choose one from hsa (default), tdc.}

\item{size}{a number specifying the minimum number of members in each annotation term to be used in the analysis. Default = 3.}
}
\value{
a list of the following components:

enrichment = enrichment results.

network = network format for \code{\link[piano:networkPlot]{piano::networkPlot()}}.

details = a list of analysis details: univsize, testMethod, pAdjusted, inputsize, numsets, minsize.
}
\description{
perform enrichment analysis using entity-level statistics e.g. p-values. The function wraps around the main functions of \pkg{\link{piano}}.
}
\details{
For pathway analysis, Metabox uses KEGG ID (e.g.C12078) for compounds, UniProt entry (e.g.P0C9J6) for proteins, NCBI-GeneID (e.g.10327) for genes.
For chemical class analysis, HMDB ID (e.g.HMDB0000001) is used for compounds.
For \code{input_data}, 1st column = list of variable/feature IDs, 2nd column = statistics and 3rd column = directions (if applicable).
}
\examples{
#out=enrichment_analyze(fnanal_data$compound_data, pcol=5, fccol=6)
#piano::networkPlot(out$network,class="non",significance=0.05)
}
\references{
Patil K. and Nielsen J. (2005) Uncovering transcriptional regulation of metabolism by using metabolic network topology. Proceedings of the National Academy of Sciences of the United States of America 102(8), 2685.

Oliveira A., Patil K., and Nielsen J. (2008) Architecture of transcriptional regulatory circuits is knitted over the topology of bio-molecular interaction networks. BMC Systems Biology 2(1), 17.

Väremo L., Nielsen J., and Nookaew I. (2013) Enriching the gene set analysis of genome-wide data by incorporating directionality of gene expression and combining statistical hypotheses and methods. Nucleic Acids Research, 41(8), pp. 4378-4391.
}
\seealso{
\pkg{\link{piano}}, \code{\link[piano:loadGSC]{piano::loadGSC()}}, \code{\link[piano:runGSA]{piano::runGSA()}}, \code{\link[piano:GSAsummaryTable]{piano::GSAsummaryTable()}}
}
\author{
Kwanjeera W \email{kwanjeera.wan@mahidol.ac.th}
}
