% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_report}
\alias{generate_report}
\title{Generate report}
\usage{
generate_report(METBObj=NULL, datsummary1=NULL, datsummary2=NULL,
note=NULL, reportfile, savedir=getwd())
}
\arguments{
\item{METBObj}{METBObj object returned from data processing functions, including \code{impute_missing_data},
\code{normalize_input_data_bydata} and \code{normalize_input_data_byqc}, \code{scale_input_data}, \code{set_input_obj}, \code{transform_input_data}.}

\item{datsummary1}{list of data returned from data analysis functions, including \code{comb_overrep_analyze}, \code{correlation_analyze},
\code{enrichment_analyze}, \code{lme_analyze}, \code{multiv_analyze}, \code{overrep_analyze}, \code{run_muvr}, \code{test_multinormality},
\code{univ_analyze}, \code{mbplsda_analyze}.}

\item{datsummary2}{list of data returned from data analysis function, including \code{test_uninormality}.}

\item{note}{text of additional information.}

\item{reportfile}{name of a report file specifically for each data processing or analysis function,
including correlation_report, impute_data_report, lme_report, mbplsda_report, multivariate_report, biomarker_report,
normalize_bydata_report, normalize_byqc_report, scale_data_report, transform_data_report and univariate_report.}

\item{savedir}{path to a directory for saving the report. Default is the current working directory.}
}
\value{
pdf file.
}
\description{
Generate report after data processing or data analysis.
}
\examples{
#sugar_dt = set_input_obj(sugar, 1,2,5)
#out=multiv_analyze(sugar_dt)
#generate_report(sugar_dt,datsummary1=out,reportfile="multivariate_report")
}
