% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_posthoc_summary}
\alias{get_posthoc_summary}
\title{Get posthoc summary}
\usage{
get_posthoc_summary(aov_res, cutoff=0.05)
}
\arguments{
\item{aov_res}{list of ANOVA results containing posthoc tests from \code{univ_analyze()}.}

\item{cutoff}{a number indicating cutoff for statistical significance (Default = 0.05).}
}
\value{
data frame of test results.
}
\description{
Get posthoc summary from ANOVA.
}
\examples{
#sugar_dt = set_input_obj(sugar, 1,2,5)
#out=univ_analyze(sugar_dt, factor2Col=3, doposthoc=TRUE)
#test_out=get_posthoc_summary(out)
}
