% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rsd}
\alias{get_rsd}
\title{Get percent RSD of QC samples}
\usage{
get_rsd(METBObj,sampleType=FALSE)
}
\arguments{
\item{METBObj}{METBObj object contains list of data.}

\item{sampleType}{a column number/index indicating sample types in the \code{METBObj$inputdata} data frame.
This parameter is required for serrf and loess methods.

Require "qc" to indicate pooled/quality control (QC) samples.}
}
\value{
data frame.
}
\description{
Calculate the relative standard deviation (RSD) of QC samples.
}
\details{
The function calculates median RSD, mean RSD and
RSD of each variable in the QC samples using RSD = sd/mean.
}
\examples{
#load('GitHub/metabox-ml/data/qc_batch_example.RData')
#out=get_rsd(qc_batch_example,sampleType=3)
}
