% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_data.R
\name{impute_missing_data}
\alias{impute_missing_data}
\title{Impute missing values}
\usage{
impute_missing_data(METBObj, method="mean", removeall=FALSE, cutoff=FALSE)
}
\arguments{
\item{METBObj}{METBObj object contains list of data.}

\item{method}{name of missing value imputation method. Choose one from the list: zero, halfmin, min, mean, median, knn, bpca, ppca, svd, rf. Default is median.}

\item{removeall}{logical indicating all variables/features with missing values are removed (TRUE) or not removed (FALSE; Default).}

\item{cutoff}{number indicating percent missing value cutoff (from 1 to 99; Default = FALSE). If cutoff = FALSE, it will impute any missing values with a chosen method.
Or else variables with missing values > percent cutoff will be removed.}
}
\value{
METBObj object.
}
\description{
Impute missing values in the input data.
}
\examples{
#out = impute_missing_data(METBObj)
}
\references{
Wolfram Stacklies, et al. pcaMethods (2007) \url{https://doi.org/10.1093/bioinformatics/btm069}.

rfImpute \url{https://www.rdocumentation.org/packages/randomForest/versions/4.6-14/topics/rfImpute}.

Marietta Kokla, et al. Random forest-based imputation (2019) \url{https://bmcbioinformatics.biomedcentral.com/articles/10.1186/s12859-019-3110-0}.
}
\author{
Kwanjeera W \email{kwanjeera.wan@mahidol.ac.th}
}
