% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_pkgs}
\alias{install_pkgs}
\title{Install package}
\usage{
install_pkgs(pkg_ls)
}
\arguments{
\item{pkg_ls}{text or list indicating the name of a package.}
}
\value{
list of unable to install packages, if any.
}
\description{
Check and install required packages.
}
\examples{
#install_pkgs("MUVR")
}
