% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{intensityplot_overview}
\alias{intensityplot_overview}
\title{Intensity plot}
\usage{
intensityplot_overview(METBObj, xCol=1, classCol=NULL,
legend_title="Group", ptsize=3, plot_title="")
}
\arguments{
\item{METBObj}{METBObj object contains list of data.}

\item{xCol}{a column number or column name indicating which variable/feature in the \code{METBObj$X} data frame to be plotted.}

\item{classCol}{a column number/index of category/factor column to show on the plot.
If not specified, \code{METBObj$classCol} is used by default.}

\item{legend_title}{text indicating the legend title.}

\item{ptsize}{a number of geom_point size.}

\item{plot_title}{text indicating plot title.}
}
\value{
ggplot object.
}
\description{
Provide intensity plot of a variable/feature
}
\examples{
#intensityplot_overview(qc_batch_example, xCol = 200, classCol = 3)
}
