% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interactionplot_byF1F2}
\alias{interactionplot_byF1F2}
\title{Interaction plot between category/factor F1 and F2}
\usage{
interactionplot_byF1F2(x_input, xCol=1, factorLv1, factorLv2, ptsize=3)
}
\arguments{
\item{x_input}{data frame of variables/features. Variables are in columns and samples are in rows.}

\item{xCol}{a column number or column name indicating which variable/feature in the \code{x_input} data frame to be plotted.}

\item{factorLv1}{a vector of F1 category/factor levels.}

\item{factorLv2}{a vector of F2 category/factor levels.}

\item{ptsize}{a number of geom_point size.}
}
\value{
ggplot object.
}
\description{
Provide interaction plot between category/factor F1 and F2.
}
\examples{
#sugar_dt = set_input_obj(sugar, 1,2,5)
#interactionplot_byF1F2(sugar_dt$X, 1, sugar_dt$Y, sugar_dt$inputdata$time)
}
