% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_gui.R
\name{launch_gui}
\alias{launch_gui}
\title{Start Shiny GUI}
\usage{
launch_gui()
}
\value{
shiny app
}
\description{
start Shiny GUI.
}
\examples{
#launch_gui()
}
\author{
Kwanjeera W \email{kwanjeera.wan@mahidol.ac.th}
}
