% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_analyze.R
\name{lme_analyze}
\alias{lme_analyze}
\title{Linear mixed-effects modeling}
\usage{
lme_analyze(METBObj, fix, random)
}
\arguments{
\item{METBObj}{METBObj object contains list of data.}

\item{fix}{a number or numeric vector indicating the fixed-effects column(s) in the \code{METBObj$inputdata} data frame.}

\item{random}{text of random-effect formular. See \code{\link[lmm2met:fitLmm]{lmm2met::fitLmm()}}.}
}
\value{
a list of the following components:

completeMod = a list of linear mixed models for all features/variables.

testRes = a list of chi-square test results for significant fixed effects.

testTable = a data frame of Intercept, coefficients and chi-square test results for significant fixed effects.

fittedDat = a data frame of fitted data.

details = a list of analysis details: testMethod, fix, random.
}
\description{
Fit linear mixed-effects models.
}
\details{
The column number or the column index always begins with 1.
Random-effect terms can be in several forms as given in Table 2 of Bates et al. (2015).
}
\examples{
#sugar_dt = set_input_obj(sugar, 1,2,5)
#out=lme_analyze(sugar_dt, fix=c(2:3), random="(1 | subjectID)")
}
\references{
Bates D, et al.,lme4 (2015). \url{https://cran.r-project.org/web/packages/lme4/vignettes/lmer.pdf}.

Wanichthanarak K, et al. lmm2met (2019) . \url{https://doi.org/10.1016/j.csbj.2019.04.009}.
}
\seealso{
\code{\link[lme4:lmer]{lme4::lmer()}}, \code{\link[stats:drop1]{stats::drop1()}}
}
\author{
Kwanjeera W \email{kwanjeera.wan@mahidol.ac.th}
}
