% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbplsda_analyze.R
\name{mbplsda_analyze}
\alias{mbplsda_analyze}
\title{Perform integrative analysis}
\usage{
mbplsda_analyze(class_data, input_data, scale=TRUE, option="none", nf=10, optdim=2,
nrepet=30, npermut=100, nboot=100, threshold=0.5, testmodel=FALSE, cpus=1)
}
\arguments{
\item{class_data}{one-column data frame of class/category/factor data.}

\item{input_data}{a list of data frames containing 2 or more data sets of explanatory/independent variables. The data frames must have the same samples and unduplicated variables.}

\item{scale}{a logical value to standardize the explanatory variables}

\item{option}{a string specifying the option for the block weighting. It can be one of none (default), uniform. If none, the block weight is equal to the block inertia.
If uniform, the weight of each explanatory data set is equal to 1/number of explanatory data sets, and the weight of the Y-block is equal to 1.}

\item{nf}{a number of principle components to be calculated. Default is 10.}

\item{optdim}{an optimal number of principle components included in the model. Default is 2.}

\item{nrepet}{a number of cross-validation repetitions. See \code{\link{testdim_mbplsda}} and \code{\link{permut_mbplsda}}.}

\item{npermut}{a number of permutations. See \code{\link{permut_mbplsda}}.}

\item{nboot}{a number of bootstrap replications. See \code{\link{boot_mbplsda}}.}

\item{threshold}{a number indicating the prediction threshold, between 0 and 1 (Default is 0.5). See \code{\link{testdim_mbplsda}}.}

\item{testmodel}{a logical variable to perform model components testing and permutation. Default is FALSE.}

\item{cpus}{a number of CPU for parallel computing.}
}
\value{
a list of the following components:

result = MBPLSDA results.

details = a list of analysis details: testMethod, nclass, inputsize, cpus, testmodel, ncomp, noptcomp, call.
}
\description{
perform integrative analysis of several data sets using a multi-block partial least squares discriminant analysis (MBPLSDA). The function wraps around the main functions of \pkg{\link{packMBPLSDA}} and \pkg{\link{ade4}}.
}
\details{
The function automatically performs all steps of MBPLSDA including generating base MBPLSDA model, determination of optimal components based on Error Rates (ER),
permutation test for model validity and parameter estimation.
}
\examples{
#sugar_dt = set_input_obj(sugar, 1,2,5)
#out = mbplsda_analyze(sugar_dt$inputdata[,2:3], list(omics = sugar_dt$X),
#nrepet=10, npermut=10, nboot=10)
#packMBPLSDA::plot_boot_mbplsda(out$result$res_boot)
#### Adipose example
#gc_dt = set_input_obj(read_input_data("data/adipose_GC.csv"),1,2,3)
#lc_pos = set_input_obj(read_input_data("data/adipose_LC_POS.csv"),1,2,3)
#lc_imp = impute_missing_data(lc_pos,method = "bpca",cutoff=30)
#lc_neg = set_input_obj(read_input_data("data/adipose_LC_NEG.csv"),1,2,3)
#out = mbplsda_analyze(data.frame(gc_dt$inputdata[,2]), list(gc = gc_dt$X,lcpos=lc_imp$X, lcneg=lc_neg$X),
#nrepet=10, npermut=3, nboot=3, testmodel = F,cpus = 2)
}
\references{
Brandolini-Bunlon M., et al. (2019) Multi-block PLS discriminant analysis. Metabolomics, 15(10):134.

Bougeard S. and Dray S. (2018) Supervised Multiblock Analysis in R with the ade4 Package.Journal of Statistical Software,86(1), 1-17.
}
\seealso{
\code{\link[packMBPLSDA:mbplsda]{packMBPLSDA::mbplsda()}}, \pkg{\link{packMBPLSDA}}, \pkg{\link{ade4}}
}
\author{
Kwanjeera W \email{kwanjeera.wan@mahidol.ac.th}
}
