% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbplsda_util.R
\name{mbplsda_plotboot_vipc}
\alias{mbplsda_plotboot_vipc}
\title{Variable importance plot}
\usage{
mbplsda_plotboot_vipc(x_input,plot_title="",ptsize=2,propbestvar=0.5)
}
\arguments{
\item{x_input}{\code{res_boot} object result from \code{mbplsda_analyze()}.}

\item{plot_title}{text indicating plot title.}

\item{ptsize}{a number of geom_point size.}

\item{propbestvar}{a number between 0 and 1 (Default = 0.5) indicating the top n-percent of variables with the best VIP values to plot.}
}
\value{
ggplot object.
}
\description{
Plot variable importance (VIP) and confidence intervals from bootstrapping simulations.
}
\details{
Modify from \code{\link[packMBPLSDA:plot_boot_mbplsda]{packMBPLSDA::plot_boot_mbplsda()}}.
}
\examples{
#sugar_dt = set_input_obj(sugar, 1,2,5)
#out = mbplsda_analyze(sugar_dt$inputdata[,2:3], list(omics = sugar_dt$X),
#nrepet=5, npermut=5, nboot=5)
#mbplsda_plotboot_vipc(out$result$res_boot)
}
\references{
Brandolini-Bunlon M., et al. (2019) Multi-block PLS discriminant analysis. Metabolomics, 15(10):134.
}
\seealso{
\code{\link[packMBPLSDA:boot_mbplsda]{packMBPLSDA::boot_mbplsda()}},\code{\link[packMBPLSDA:plot_boot_mbplsda]{packMBPLSDA::plot_boot_mbplsda()}}
}
\author{
Kwanjeera W \email{kwanjeera.wan@mahidol.ac.th}
}
