% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbplsda_util.R
\name{mbplsda_plotpermut}
\alias{mbplsda_plotpermut}
\title{Permutation testing plot}
\usage{
mbplsda_plotpermut(x_input,plot_title="",ptsize=3)
}
\arguments{
\item{x_input}{\code{res_permut} object result from \code{mbplsda_analyze()}.}

\item{plot_title}{text indicating plot title.}

\item{ptsize}{a number of geom_point size.}
}
\value{
ggplot object.
}
\description{
Plot prediction error rate against the percentage of modified Y-block values in permutation testing.
}
\details{
Modify from \code{\link[packMBPLSDA:plot_permut_mbplsda]{packMBPLSDA::plot_permut_mbplsda()}}.
}
\examples{
#sugar_dt = set_input_obj(sugar, 1,2,5)
#out = mbplsda_analyze(sugar_dt$inputdata[,2:3], list(omics = sugar_dt$X),
#nrepet=5, npermut=5, nboot=5)
#mbplsda_plotpermut(out$result$res_permut)
}
\references{
Brandolini-Bunlon M., et al. (2019) Multi-block PLS discriminant analysis. Metabolomics, 15(10):134.
}
\seealso{
\code{\link[packMBPLSDA:permut_mbplsda]{packMBPLSDA::permut_mbplsda()}},\code{\link[packMBPLSDA:plot_permut_mbplsda]{packMBPLSDA::plot_permut_mbplsda()}}
}
\author{
Kwanjeera W \email{kwanjeera.wan@mahidol.ac.th}
}
