% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbplsda_util.R
\name{mbplsda_plottestdim}
\alias{mbplsda_plottestdim}
\title{Prediction error rate plot}
\usage{
mbplsda_plottestdim(x_input,plotvalida=TRUE,ptsize=2)
}
\arguments{
\item{x_input}{\code{res_optimal} object result from \code{mbplsda_analyze()}.}

\item{plotvalida}{a logical variable to plot validation (TRUE) or calibration subset (FALSE).}

\item{ptsize}{a number of geom_point size.}
}
\value{
ggplot object.
}
\description{
Plot prediction error rate against the corresponding principle component number.
}
\details{
Modify from \code{\link[packMBPLSDA:plot_testdim_mbplsda]{packMBPLSDA::plot_testdim_mbplsda()}}.
}
\examples{
#sugar_dt = set_input_obj(sugar, 1,2,5)
#out = mbplsda_analyze(sugar_dt$inputdata[,2:3], list(omics = sugar_dt$X),
#nrepet=5, npermut=5, nboot=5)
#mbplsda_plottestdim(out$result$res_optimal)
}
\references{
Brandolini-Bunlon M., et al. (2019) Multi-block PLS discriminant analysis. Metabolomics, 15(10):134.
}
\seealso{
\code{\link[packMBPLSDA:testdim_mbplsda]{packMBPLSDA::testdim_mbplsda()}},\code{\link[packMBPLSDA:plot_testdim_mbplsda]{packMBPLSDA::plot_testdim_mbplsda()}}
}
\author{
Kwanjeera W \email{kwanjeera.wan@mahidol.ac.th}
}
