% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbplsda_util.R
\name{mbplsda_screeplot}
\alias{mbplsda_screeplot}
\title{Scree plot}
\usage{
mbplsda_screeplot(x_input,plot_title="",ptsize=2)
}
\arguments{
\item{x_input}{numeric vectors of eigenvalues.}

\item{plot_title}{text indicating plot title.}

\item{ptsize}{a number of geom_point size.}
}
\value{
ggplot object.
}
\description{
Plot eigenvalues against the corresponding principle component number.
}
\examples{
#sugar_dt = set_input_obj(sugar, 1,2,5)
#out = mbplsda_analyze(sugar_dt$inputdata[,2:3], list(omics = sugar_dt$X),
#nrepet=5, npermut=5, nboot=5)
#mbplsda_screeplot(out$result$base_model$eig)
}
\author{
Kwanjeera W \email{kwanjeera.wan@mahidol.ac.th}
}
