% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multiv_loadingplot}
\alias{multiv_loadingplot}
\title{Loading plot}
\usage{
multiv_loadingplot(loading_data, oloading_data=NULL, ptsize=2, plot_title="")
}
\arguments{
\item{loading_data}{a numerical matrix or a data frame of x loadings, see details.}

\item{oloading_data}{a numerical matrix or a data frame of orthogonal loadings, for OPLS only. Default is \code{NULL}.}

\item{ptsize}{a number of geom_point size.}

\item{plot_title}{text indicating plot title.}
}
\value{
ggplot object.
}
\description{
Provide loading plot for multivariate analysis.
}
\details{
A two-column numerical matrix or a two-column data frame of x loadings is required for PCA and PLSDA.
}
\examples{
#adipose_dt = set_input_obj(adipose,2,3,4)
#out=multiv_analyze(adipose_dt, method="pls", scale="standard")
#multiv_loadingplot(out$loading_val)
}
