% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multiv_loadingplot_bypc}
\alias{multiv_loadingplot_bypc}
\title{Loading plot by PC}
\usage{
multiv_loadingplot_bypc(loading_data, pc=1, plot_title="")
}
\arguments{
\item{loading_data}{a numerical matrix or a data frame of x loadings or orthogonal loadings.}

\item{pc}{a number indicating a principal component to plot. 1st PC is plotted by default.}

\item{plot_title}{text indicating plot title.}
}
\value{
ggplot object.
}
\description{
Provide loading plot by a principal component.
}
\examples{
#adipose_dt = set_input_obj(adipose,2,3,4)
#out=multiv_analyze(adipose_dt, method="pls", scale="standard")
#multiv_loadingplot_bypc(out$loading_val)
}
