% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multiv_scoreplot}
\alias{multiv_scoreplot}
\title{Score plot}
\usage{
multiv_scoreplot(METBObj,score_data,pcx=0,pcy=0,oscore_data=NULL,
legend_title="Group", ptsize=2, plot_title="")
}
\arguments{
\item{METBObj}{METBObj object contains list of data.}

\item{score_data}{a numerical matrix or a data frame of x scores, see details.}

\item{pcx}{a number indicating variance explained by a principal component (x-axis).}

\item{pcy}{a number indicating variance explained by a principal component (y-axis).}

\item{oscore_data}{a numerical matrix or a data frame of orthogonal scores, for OPLS only. Default is \code{NULL}.}

\item{legend_title}{text indicating the legend title.}

\item{ptsize}{a number of geom_point size.}

\item{plot_title}{text indicating plot title.}
}
\value{
ggplot object.
}
\description{
Provide score plot for multivariate analysis.
}
\details{
A two-column numerical matrix or a two-column data frame of x scores is required for PCA and PLSDA.
}
\examples{
#adipose_dt = set_input_obj(adipose,2,3,4)
#out=multiv_analyze(adipose_dt, method="opls", scale="standard")
#multiv_scoreplot(adipose_dt,out$score_val,pcx=out$model_summary$R2X[1],
#pcy=out$model_summary$R2X[2],out$oscore_val)
}
