% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multiv_viploadingplot}
\alias{multiv_viploadingplot}
\title{VIP and loading plot}
\usage{
multiv_viploadingplot(vip_data, loading_data, oloading_data=NULL, plot_title="")
}
\arguments{
\item{vip_data}{a numerical vector of VIP.}

\item{loading_data}{a numerical matrix or a data frame of x loadings, see details.}

\item{oloading_data}{a numerical matrix or a data frame of orthogonal loadings, for OPLS only. Default is \code{NULL}.}

\item{plot_title}{text indicating plot title.}
}
\value{
ggplot object.
}
\description{
Provide combined VIP and loading plot for multivariate analysis.
}
\examples{
#adipose_dt = set_input_obj(adipose,2,3,4)
#out=multiv_analyze(adipose_dt, method="pls", scale="standard")
#multiv_viploadingplot(out$vip_val,out$loading_data)
}
