% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multiv_vipplot}
\alias{multiv_vipplot}
\title{VIP plot}
\usage{
multiv_vipplot(vip_data, plot_title="")
}
\arguments{
\item{vip_data}{a numerical vector of VIP.}

\item{plot_title}{text indicating plot title.}
}
\value{
ggplot object.
}
\description{
Provide VIP plot for multivariate analysis.
}
\examples{
#adipose_dt = set_input_obj(adipose,2,3,4)
#out=multiv_analyze(adipose_dt, method="pls", scale="standard")
#multiv_vipplot(out$vip_val)
}
