% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUVR_util.R
\name{MUVR_getvip}
\alias{MUVR_getvip}
\title{Get VIP ranks}
\usage{
MUVR_getvip(MVObj, model="min")
}
\arguments{
\item{MVObj}{MVObj object contains list of data.}

\item{model}{name indicating Which model to use ("min" [default], "mid", or "max").}
}
\value{
data frame with order, name and average rank of features/variables.
}
\description{
Get VIP ranks from MUVR object.
}
\details{
Overriding \code{\link[MUVR:getVIP]{MUVR::getVIP()}}.
}
\examples{
#res_muvr = run_muvr(METBObj, method="PLS")
#MUVR_getvip(res_muvr)
}
\references{
Lin Shi, et al. MUVR (2019) \url{https://doi.org/10.1093/bioinformatics/bty710}.
}
\seealso{
\code{\link[MUVR:getVIP]{MUVR::getVIP()}}
}
\author{
Kwanjeera W \email{kwanjeera.wan@mahidol.ac.th}
}
