% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUVR_util.R
\name{MUVR_plotvip}
\alias{MUVR_plotvip}
\title{VIP rank plot}
\usage{
MUVR_plotvip(MVObj,model="min",plot_title="")
}
\arguments{
\item{MVObj}{MVObj object contains list of data.}

\item{model}{name indicating Which model to use ("min" [default], "mid", or "max").}

\item{plot_title}{text indicating plot title.}
}
\value{
ggplot object
}
\description{
Plot VIP ranking in MUVR object.
}
\details{
Overriding \code{\link[MUVR:plotVIP]{MUVR::plotVIP()}}.
}
\examples{
#res_muvr = run_muvr(METBObj, method="PLS")
#MUVR_plotvip(res_muvr)
}
\references{
Lin Shi, et al. MUVR (2019) \url{https://doi.org/10.1093/bioinformatics/bty710}.
}
\seealso{
\code{\link[MUVR:plotVIP]{MUVR::plotVIP()}}
}
\author{
Kwanjeera W \email{kwanjeera.wan@mahidol.ac.th}
}
