% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_input_data_bydata.R
\name{normalize_input_data_bydata}
\alias{normalize_input_data_bydata}
\title{Data-driven normalizations}
\usage{
normalize_input_data_bydata(METBObj, method="median", ref=NULL)
}
\arguments{
\item{METBObj}{METBObj object contains list of data.}

\item{method}{name of normalization method. Choose one from the list: contrast, cubic, eigenms, linear, loess, mean, median, pqn, quantile, sum, vsn. Default is median.}

\item{ref}{a row number/index or a vector of row numbers of reference/control sample(s) in the \code{METBObj$inputdata} data frame.
This parameter is required for pqn method.}
}
\value{
METBObj object.
}
\description{
Normalize samples by data-driven normalization methods.
}
\details{
The row number or the row index always begins with 1.
}
\examples{
#out = normalize_input_data_bydata(METBObj)
}
\references{
Astrand M. contrast (2003) \url{https://doi.org/10.1089/106652703763255697}.

Frank Dieterle, et al. pqn (2006) \url{https://pubs.acs.org/doi/10.1021/ac051632c}.

Bolstad Ben. preprocessCore.quantiles (2021) \url{https://github.com/bmbolstad/preprocessCore}.

Yuliya V. Karpievitch, et al. EigenMS (2014) \url{https://doi.org/10.1371/journal.pone.0116221}.

Bethanne M.Warrack, et al. MSTUS (2009) \url{https://doi.org/10.1016/j.jchromb.2009.01.007}.

Sandrine Dudoit, et al. cyclic loess (2002) \url{https://wwwf.imperial.ac.uk/~das01/BioinformaticsMSc/Papers/sinica.final.pdf}.

Christopher Workman, et al. cubic spline (2002) \url{https://genomebiology.biomedcentral.com/articles/10.1186/gb-2002-3-9-research0048}.

Chetwynd, AJ, et al. MSTUS (2016) \url{https://doi.org/10.1016/j.chroma.2015.12.056}.

Wolfgang Huber, et al. VSN (2002) \url{https://doi.org/10.1093/bioinformatics/18.suppl_1.S96}.

E. S. Motakis, et al. Variance stabilization and normalization (2006) \url{https://doi.org/10.1093/bioinformatics/btl412}.
}
\seealso{
\code{\link[preprocessCore:normalize.quantiles]{preprocessCore::normalize.quantiles()}}, \code{\link[affy:normalize.loess]{affy::normalize.loess()}}, \code{\link[affy:normalize.qspline]{affy::normalize.qspline()}}, \code{\link[affy:normalize.invariantset]{affy::normalize.invariantset()}}, \code{\link[vsn:vsn2]{vsn::vsn2()}}
}
\author{
Kwanjeera W \email{kwanjeera.wan@mahidol.ac.th}
}
