% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pcaplot_overview}
\alias{pcaplot_overview}
\title{PCA plot}
\usage{
pcaplot_overview(METBObj, classCol=NULL, shapeCol=NULL, px=1, py=2,
scale=FALSE, legend_title="Group", ptsize=3, plot_title="")
}
\arguments{
\item{METBObj}{METBObj object contains list of data.}

\item{classCol}{a column number/index of category/factor column to show on the plot.
If not specified, \code{METBObj$classCol} is used by default.}

\item{shapeCol}{a column number/index of category/factor column to show as different point shapes on the plot.
If not specified, a filled circle is used by default.}

\item{px}{a number indicating a principal component to be plotted on x-axis (default=1).}

\item{py}{a number indicating a principal component to be plotted on y-axis (default=2).}

\item{scale}{a logical value indicating whether the variables should be scaled to have unit variance before the analysis.
Default is FALSE.}

\item{legend_title}{text indicating the legend title.}

\item{ptsize}{a number of geom_point size.}

\item{plot_title}{text indicating plot title.}
}
\value{
ggplot object.
}
\description{
Provide PCA plot overview by category/factor.
}
\examples{
#pcaplot_overview(METBObj)
}
