% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pvalplot_overview}
\alias{pvalplot_overview}
\title{p-value plot}
\usage{
pvalplot_overview(x_input, cutoff=0.05,
plot_title="", ptsize=2)
}
\arguments{
\item{x_input}{data frame of p-values/adjusted p-values. A data frame contains only one column using variables as \code{row.names}.}

\item{cutoff}{a number indicating statistical significance.}

\item{plot_title}{text indicating the plot title.}

\item{ptsize}{a number of geom_point size.}
}
\value{
ggplot object.
}
\description{
Provide p-value plot overview.
}
\examples{
#sugar_dt = set_input_obj(sugar, 1,2,5)
#uni_out=univ_analyze(sugar_dt)
#pvalplot_overview(data.frame(uni_out$p_adj),plot_title = "Adjusted p-value")
}
