% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rlaplot_overview}
\alias{rlaplot_overview}
\title{RLA plot}
\usage{
rlaplot_overview(METBObj, classCol=NULL, type="wg", dolog=TRUE,
doavg=FALSE, limitx=FALSE, plot_title="")
}
\arguments{
\item{METBObj}{METBObj object contains list of data.}

\item{classCol}{a column number/index of category/factor column to show on the plot.
If not specified, \code{METBObj$classCol} is used by default.}

\item{type}{text indicating whether within group ("wg") or across group ("ag") to be plotted.}

\item{dolog}{a logical value indicating whether the variables should be log2-transformed prior to plotting.
Default is TRUE.}

\item{doavg}{a logical value indicating whether the standardized variables should be obtained by
removing the mean from each variable. Default is FALSE.}

\item{limitx}{a logical value to limit no. of samples on the plot.}

\item{plot_title}{text indicating plot title.}
}
\value{
ggplot object.
}
\description{
Provide RLA plot overview to visualize sample variation.
}
\examples{
#sugar_dt = set_input_obj(sugar, 1,2,5)
#rlaplot_overview(sugar_dt)
}
\references{
Rlaplots \url{https://rdrr.io/cran/metabolomics/man/RlaPlots.html}.
}
