% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_muvr.R
\name{run_muvr}
\alias{run_muvr}
\title{Run MUVR}
\usage{
run_muvr(METBObj, scale=FALSE, nRep=3, varRatio=0.75, partitionSize=0.7,
method="PLS", DA=TRUE, fitness="AUROC")
}
\arguments{
\item{METBObj}{METBObj object contains list of data.}

\item{scale}{a logical variable indicating whether the variables should be scaled to have unit variance before PLS modelling. Default is FALSE.}

\item{nRep}{number of repetitions of double CV.}

\item{varRatio}{ratio of variables to include in inner loop iterations. Default is 0.75.}

\item{partitionSize}{number of data holding for training set. Default is 0.7.}

\item{method}{ML method. Choose one from the list: PLS, RF. Default is PLS.}

\item{DA}{a logical variable for performing classification. Default is TRUE. For performing regression, DA=FALSE.}

\item{fitness}{fitness function for model tuning and evaluation. Choose AUROC (default) or MISS for classification; or choose RMSEP for regression.}
}
\value{
MVObject object.
}
\description{
Perform ML analysis using MUVR package.
}
\details{
Based on MUVR version 0.0.976. (2023-03-09). The \code{partitionSize} is for calculating \code{nOuter}.
}
\examples{
#res_muvr = run_muvr(METBObj, method="PLS")
#MUVR::plotVAL(res_muvr) #need more distinct colors
#MUVR::plotMV(res_muvr, model = 'min')
#MUVR::plotStability(res_muvr, model = 'min')
#MUVR::plotVIP(res_muvr, model = 'min') #error if nrep=1
#MUVR::getVIP(res_muvr, model = 'min')
}
\references{
Lin Shi, et al. MUVR (2019) \url{https://doi.org/10.1093/bioinformatics/bty710}.
}
\seealso{
\code{\link[MUVR:MUVR]{MUVR::MUVR()}}, \code{\link[MUVR:vectSamp]{MUVR::vectSamp()}}
}
\author{
Kwanjeera W \email{kwanjeera.wan@mahidol.ac.th}
}
