% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_input_data.R
\name{scale_input_data}
\alias{scale_input_data}
\title{Scale data}
\usage{
scale_input_data(METBObj, method="pareto")
}
\arguments{
\item{METBObj}{METBObj object contains list of data.}

\item{method}{name of scaling method. Choose one from the list: auto, level, pareto, power, range, vast. Default is pareto}
}
\value{
METBObj object
}
\description{
Scale data for between-variable/feature variations
}
\examples{
#out = scale_input_data(METBObj)
}
\references{
Qingxia Yang, et al. NOREVA (2020) \url{https://doi.org/10.1093/bib/bbz137}.
}
\author{
Kwanjeera W \email{kwanjeera.wan@mahidol.ac.th}
}
