% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_input_obj.R
\name{set_input_obj}
\alias{set_input_obj}
\title{Set input data object}
\usage{
set_input_obj(inputdata, idCol, classCol, xCol)
}
\arguments{
\item{inputdata}{dataframe of input data. Samples are in rows and variables/features are in columns.}

\item{idCol}{a column number/index of sample ID column which indicates independent or repeated sample structure of the experimental design. Default is 1.}

\item{classCol}{a column number/index of category/factor column (for hypothesis testing and classification)
or response column (for regression). Default is 2.}

\item{xCol}{a column number/index of the first variable/feature column. Default is 3.}
}
\value{
METBObj object; a list of the following components:

inputdata = data frame of input data.

ID = vector of all ids.

orgID = vector of original ids.

unik = vector of logical variables indicating unique ids, for MUVR.

unikID = vector of unique ids, for MUVR.

idCol = a column number/index indicates independent or repeated sample structure of the experimental design.

xCol = a column number/index of the first variable/feature column in the inputdata data frame.

classCol = a column number/index of category/factor column (for hypothesis testing and classification)
or response column (for regression) in the inputdata data frame.

X = data frame of variables/features.

Y = vector of factors (for classification) or response variables (for regression).

isRepeated = a logical variable indicating a repeated-measures design.

details = a list of object details.
}
\description{
Set input data object
}
\examples{
#out = set_input_obj(read_input_data('filename'),1,2,3)
#out = set_input_obj(adipose, 2,3,4)
}
\author{
Kwanjeera W \email{kwanjeera.wan@mahidol.ac.th}
}
