% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_multiequalvar}
\alias{test_multiequalvar}
\title{Box's M-test for homogeneity of multivariate data}
\usage{
test_multiequalvar(METBObj)
}
\arguments{
\item{METBObj}{METBObj object contains list of data.}
}
\value{
list of test results.
}
\description{
Perform Box's M-test on all variables.
}
\examples{
#test_multiequalvar(METBObj) #Not exported function
}
