% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_multinormality}
\alias{test_multinormality}
\title{Shapiro-Wilk test of multivariate normality}
\usage{
test_multinormality(METBObj)
}
\arguments{
\item{METBObj}{METBObj object contains list of data.}
}
\value{
list of test results.
}
\description{
Perform Shapiro-Wilk test on all variables.
}
\examples{
#test_multinormality(METBObj) #Not exported function
}
