% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_uniequalvar}
\alias{test_uniequalvar}
\title{Levene's test for homogeneity of variance across groups}
\usage{
test_uniequalvar(METBObj)
}
\arguments{
\item{METBObj}{METBObj object contains list of data.}
}
\value{
list of test results.
}
\description{
Perform Levene's test on each variable.
}
\examples{
#test_uniequalvar(METBObj)
}
