% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_uninormality}
\alias{test_uninormality}
\title{Shapiro-Wilk test of univariate normality}
\usage{
test_uninormality(METBObj)
}
\arguments{
\item{METBObj}{METBObj object contains list of data.}
}
\value{
list of test results.
}
\description{
Perform Shapiro-Wilk test on each variable.
}
\examples{
#test_uninormality(METBObj)
}
