% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_input_data.R
\name{transform_input_data}
\alias{transform_input_data}
\title{Transform data}
\usage{
transform_input_data(METBObj, method="sqrt")
}
\arguments{
\item{METBObj}{METBObj object contains list of data.}

\item{method}{name of scaling method. Choose one from the list: cube, glog10, glog2, log10, log2, sqrt. Default is sqrt}
}
\value{
METBObj object.
}
\description{
Transform variables/features to be closer to normal or Gaussian distribution. The generalized logarithm transformation is used to stabilize variance.
}
\examples{
#out = transform_input_data(METBObj)
}
\references{
David M. Rocke, Blythe Durbin. glog (2003) \url{https://doi.org/10.1093/bioinformatics/btg107}.

Wolfgang Huber, et al. VSN (2002) \url{https://doi.org/10.1093/bioinformatics/18.suppl_1.S96}.

E. S. Motakis, et al. Variance stabilization and normalization (2006) \url{https://doi.org/10.1093/bioinformatics/btl412}.
}
\author{
Kwanjeera W \email{kwanjeera.wan@mahidol.ac.th}
}
