% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univ_analyze.R
\name{univ_analyze}
\alias{univ_analyze}
\title{Univariate analysis}
\usage{
univ_analyze(METBObj, var.equal = FALSE, ispara = FALSE,
factor2Col = FALSE, doposthoc = FALSE)
}
\arguments{
\item{METBObj}{METBObj object contains list of data.}

\item{var.equal}{a logical variable for pairwise comparisons indicating
whether to treat the two variances as being equal (see \code{\link[stats:t.test]{stats::t.test()}}).}

\item{ispara}{a logical variable wheter to perform parametric (TRUE) or non-parametric (FALSE) test. Default is FALSE.}

\item{factor2Col}{a column number/index of the second factor in the \code{METBObj$inputdata} data frame.
This parameter is used in Two-way ANOVA.}

\item{doposthoc}{a logical variable wheter to perform posthoc test for One-way ANOVA and Two-way ANOVA. Default is FALSE.}
}
\value{
a list of the following components:

stat_summ = a data frame of statistical summary including median, mean, SD (sd) and fold change (fc).
If 2 levels, fc=level1/level2, if more than 2 levels, fc=each_level/all_mean. Levels are ordered by characters.

p_value = a numeric vector or matrix of p-values.

p_adj = a numeric vector or matrix of adjusted p-values.

posthoc_data = a list of post-hoc test results. Return empty list for Pairwise comparisons or \code{doposthoc = FALSE}.

posthoc_table = a data frame of post-hoc test results. Return empty data frame for Pairwise comparisons or \code{doposthoc = FALSE}.

details = a list of analysis details: isParametric, testMethod, pAdjusted, posthocTest.
}
\description{
Perform univariate analysis.
}
\details{
The column number or the column index always begins with 1.
Statistical analysis for independent samples will be performed for unequal sample sizes.
The friedman.test is the non-parametric alternative to the repeated measures one-way ANOVA and
it requires an unreplicated complete block design.
}
\examples{
#sugar_dt = set_input_obj(sugar, 1,2,5)
#out=univ_analyze(sugar_dt, factor2Col=3)
}
\references{
Dunn, O.J. 1964. Multiple comparisons using rank sums. Technometrics 6:241-252.

Sokal, R.R. and F.J. Rohlf. 1995. scheirerRayHare. Biometry. 3rd ed. W.H. Freeman, New York.

http://rcompanion.org/handbook/F_14.html.
}
\seealso{
\code{\link[stats:t.test]{stats::t.test()}}, \code{\link[stats:wilcox.test]{stats::wilcox.test()}}
}
\author{
Kwanjeera W \email{kwanjeera.wan@mahidol.ac.th}
}
