% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_databox.R
\docType{data}
\name{anno_databox}
\alias{anno_databox}
\title{Annotation data}
\format{
list of meta data and pair data for organisms:
\describe{
Gene downloaded from Ensembl Human genes (GRCh38.p14).
Protein downloaded from UniProt release 2024_05.
Pathway downloaded from KEGG release 112.0.
}
}
\source{
\url{""}
}
\usage{
anno_databox
}
\description{
Annotation data for overrepresentation and enrichment analysis
}
\keyword{datasets}
